Moving routing configuration to AppRoutingModule
